Custom object interface storage instead of wl_map
Previously, we used wl_map to store a mapping of object IDs to
interfaces.  But this only works if we map every ID in order with no
gaps.  Occasionally we miss a new ID creation (eg. if we're missing a
protocol) and then the IDs in the map get out of sync and so we fail to
lookup IDs later and it all goes horribly wrong.

Instead just make a simple array for each of client/server IDs and store
the interfaces in there.  This still works if we miss an ID since we're
just storing / looking up by ID so holes are ok.