Add optional dependency on percetto/perfetto
Perfetto is a tracing system. I'd like to add an option for
wayland-tracer to add events to it so you can see when events (like
keyboard/mouse input or frame callbacks) occur in the trace.

Add a dependency on percetto, which is a C wrapper for the C++ perfetto
SDK.  These are pulled in automatically as subprojects.

This can be disabled by running meson with `-Dperfetto=disabled`.